feat(iife): group client iife
This implements an immediatetely invoked function execution to be
handled by the browser to prevent page reloads when a user is attempting
to execute a command from the browser. Since this uses native browser
events to submit a form request to the backend, this caters for
preventing the default behaviour which allows the client not to
experience page reloads for example when running a command such as
increaseing the volume which has multiple keystrokes.